Bug fix: edge cases where ids contain prefixes
Some genesets contain IDs of different type from the other IDSs on the
set, and are prefixed by a string with a pipe. e.g. 'ensg|' for ensembl.

This is dealt with by removing prefixes from IDs that contain it, and
making sure that all gene tuples are size 3.

Also, if a geneset has mixed ID types, we fall back to searching all
scopes.