Upgrade plugins to use new function to get and format query results (#54)
* Ignore test data dirs

* Fix bug: Calling get_results with a list of str

Calling get_results with a list of gene ids (instead of gene tuples) as
input parameter would cause an index not found error. Now the input list
is converted to tuples if it contains strings.

* Update to new gene_utils functions

- Use new lookup_results function to get gene query results.
- Add new taxid to dictionary, and output error logging if a taxid is
not found.

* Update mapping

* Use new get_results function

* Add metabolites count to docs

* Update mapping

* Rename 'dups' field to 'ids'

* Use new get_results function

* Update mapping

* Unlist documents

* Use new geneset utils workflow

* Update genmap2 url

* Update license url

* Add type checks

Make sure all inputs are strings for .

Make sure all inputs are tuples of the same length.

* Update mapping

* Use mygeneset_utils functions

* Update mapping

* Formatting changes

* Comment out unused variable

* Restructure module

* Feat: Homolog gene conversions Fix: dup genesets

New function `query_mygene_homologs()` allows creating genesets from a
list of genes containing taxids from various organisms.

Fix a bug where it was possible to create genesets containing duplicate
genes. If, for example, the list contained two different types of ids
belonging to the same gene, each would be a sepparate list entry with
the same mygene_id.

* Update parser to use new homolog lookup function

* Change geneset utilities module name

* Fix typo

* Update go to use new geneset utils