Bug fix: Clear gene list on 'replace' update.
If gene list is empty during a replace operation i.e. all of the genes
passed are 'not found' , the 'genes' field should be replaced with an
empty list. Before, if no genes were found, the document would be
missing a 'genes' field, and no change was made because any missing
field in the original document is not overwritten.